Merge branch '1GbE' of git://git.kernel.org/pub/scm/linux/kernel/git/jkirsher/next-queue
Jeff Kirsher says:

====================
1GbE Intel Wired LAN Driver Updates 2016-09-27

This series contains updates to igb and igbvf.

Wei Yongjun makes a function static to shut up sparse.

Todd bumps the igb and igbvf version, which is long overdue.

Jake fixes an issue where the PPS SYS_WRAP interrupt was not re-enabled
after a reset, which resulted in disabling of the PPS signaling.

v2: dropped patch 5 of the original series, since the PCI quirk patch
    needs to be reworked by Alex and Sasha to address issues that Bjorn
    Helgaas and Alex Williamson brought up.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: dfdc771d03c13d0e170c27cd845bdab716e73576 ]